separated profile routes to worksamples, basic, instruments, and instrument js files
